+++ recorder.py
@@ -14,11 +14,6 @@ from struct import pack
 import pyaudio
 import wave
 
-CHUNK_SIZE = 1024
-FORMAT = pyaudio.paInt16
-RATE = 44100
-
-
 def normalize(sndData):
     "Average the volume out"
     MAXIMUM = 16384
@@ -32,9 +27,9 @@ def normalize(sndData):
 
 def add_silence(sndData, seconds):
     "Add silence to the start and end of 'sndData' of length 'seconds' (float)"
-    r = array('h', [0 for i in range(int(seconds * RATE))])
+    r = array('h', [0 for i in range(int(seconds * shared.setting['sample_rate']))])
     r.extend(sndData)
-    r.extend([0 for i in range(int(seconds * RATE))])
+    r.extend([0 for i in range(int(seconds * shared.setting['sample_rate']))])
     return r
 
 
@@ -84,7 +79,7 @@ def trim(sndData, threshold):
     return sndData
 
 
-def environment_noise(samplingTime):
+def environment_noise(samplingTime, chunk=512):
     """
     Record the sound in a certain duration as the environment noise, and calcuate its power.
 
@@ -98,16 +93,16 @@ def environment_noise(samplingTime):
     """
     def calc_threshold():
         p = pyaudio.PyAudio()
-        stream = p.open(format=FORMAT, channels=1, rate=RATE,
+        stream = p.open(format=pyaudio.paInt16, channels=1, rate=shared.setting['sample_rate'],
                         input=True, output=True,
-                        frames_per_buffer=CHUNK_SIZE)
+                        frames_per_buffer=chunk)
 
         r = array('h')
 
         noise = []
-        for i in range(int(samplingTime * RATE // CHUNK_SIZE)):
+        for i in range(int(samplingTime * shared.setting['sample_rate'] // chunk)):
             # little endian, signed short
-            sndData = array('h', stream.read(CHUNK_SIZE))
+            sndData = array('h', stream.read(chunk))
             if byteorder == 'big':
                 sndData.byteswap()
             r.extend(sndData)
@@ -123,7 +118,7 @@ def environment_noise(samplingTime):
     return threshold
 
 
-def recordSound(threshold, minRecordTime=0, maxSoundLength=60 * RATE, feedback=False):
+def recordSound(threshold, minRecordTime=0, maxSoundLength=600*44100, feedback=False, chunk=512):
     """
     Record sound from the microphone and return the data as an array of signed shorts.
 
@@ -136,14 +131,14 @@ def recordSound(threshold, minRecordTime=0, maxSoundLength=60 * RATE, feedback=F
     todo
     """
     p = pyaudio.PyAudio()
-    stream = p.open(format=FORMAT, channels=1, rate=RATE,
+    stream = p.open(format=pyaudio.paInt16, channels=1, rate=shared.setting['sample_rate'],
                     input=True, output=True,
-                    frames_per_buffer=CHUNK_SIZE)
+                    frames_per_buffer=chunk)
 
     num_silent = 0
     speech_threshold = 4 * threshold
     num_sound = 0
-    silent_limit = 1 * RATE
+    silent_limit = 1 * shared.setting['sample_rate']
 
     onset = 0
     onset_detected = False
@@ -151,9 +146,9 @@ def recordSound(threshold, minRecordTime=0, maxSoundLength=60 * RATE, feedback=F
     r = array('h')
     while num_sound < maxSoundLength and len(r) < (minRecordTime + maxSoundLength) and num_silent < silent_limit:
         # little endian, signed short
-        s = stream.read(CHUNK_SIZE)
+        s = stream.read(chunk)
         if feedback:
-            stream.write(s, CHUNK_SIZE)
+            stream.write(s, chunk)
 
         sndData = array('h', s)
         if byteorder == 'big':
@@ -177,13 +172,13 @@ def recordSound(threshold, minRecordTime=0, maxSoundLength=60 * RATE, feedback=F
                 num_silent = 0
         else:
             if max(sndData) > speech_threshold:
-                for idx in range(1, len(r), CHUNK_SIZE):
-                    if np.max(r[-idx - CHUNK_SIZE:-idx]) < threshold:
-                        onset = len(r) - idx - CHUNK_SIZE
+                for idx in range(1, len(r), chunk):
+                    if np.max(r[-idx - chunk:-idx]) < threshold:
+                        onset = len(r) - idx - chunk
                         onset_detected = True
                         break
 
-    sample_width = p.get_sample_size(FORMAT)
+    sample_width = p.get_sample_size(pyaudio.paInt16)
     stream.stop_stream()
     stream.close()
     p.terminate()
@@ -217,7 +212,7 @@ def recordSound_tofile(path, filename, threshold, min_record_duration, maxSoundL
         wf = wave.open(path + '/' + filename + '.wav', 'wb')
         wf.setnchannels(1)
         wf.setsampwidth(sample_width)
-        wf.setframerate(RATE)
+        wf.setframerate(shared.setting['sample_rate'])
         wf.writeframes(data)
         wf.close()
         return True
@@ -225,68 +220,10 @@ def recordSound_tofile(path, filename, threshold, min_record_duration, maxSoundL
         return False
 
 
-def recordSound_testing(threshold, RecordTime=[0,60], maxSoundLength=60 * RATE, feedback=False):
+def recordSound_testing(threshold, recordingLength=[0,60], soundLength=[], trim='both', feedback=False):
     """
-    Record a word or words from the microphone and 
-    return the data as an array of signed shorts.
+    何时停止： 录音时长 (min,max), 声音时长 (min,max), trim='both','left','right'
+    优先级（冲突以短的来）？固定时间？应用场景？
     """
-    RecordTime = RecordTime[0]*,RecordTime[1]*shared.setting['sample_rate']
-    p = pyaudio.PyAudio()
-    stream = p.open(format=FORMAT, channels=1, rate=RATE,
-                    input=True, output=True,
-                    frames_per_buffer=CHUNK_SIZE)
-
-    num_silent = 0
-    speech_threshold = 4 * threshold
-    num_sound = 0
-    silent_limit = 1 * RATE
-
-    onset = 0
-    onset_detected = False
-
-    r = array('h')
-    while num_sound < maxSoundLength and len(r) < (minRecordTime + maxSoundLength) and num_silent < silent_limit:
-        # little endian, signed short
-        s = stream.read(CHUNK_SIZE)
-        if feedback:
-            stream.write(s, CHUNK_SIZE)
-
-        sndData = array('h', s)
-        if byteorder == 'big':
-            sndData.byteswap()
-        r.extend(sndData)
-
-        # for e in shared.pg.event.get():
-        #     if e.type == KEYDOWN:
-        #         k = e.key
-        #         if k == 27:
-        #             shared.pg.quit()
-        #         elif k == K_F12:
-        #             suspend()
-
-        if onset_detected:
-            num_sound += len(sndData)
-
-            if is_silent(sndData, threshold):
-                num_silent += len(sndData)
-            else:
-                num_silent = 0
-        else:
-            if max(sndData) > speech_threshold:
-                for idx in range(1, len(r), CHUNK_SIZE):
-                    if np.max(r[-idx - CHUNK_SIZE:-idx]) < threshold:
-                        onset = len(r) - idx - CHUNK_SIZE
-                        onset_detected = True
-                        break
-
-    sample_width = p.get_sample_size(FORMAT)
-    stream.stop_stream()
-    stream.close()
-    p.terminate()
-
-    r = trim(r, threshold)
-    # r = normalize(r)
-    r = r[:maxSoundLength]
-    r = np.require(np.tile(r, (2, 1)).T, requirements='C')
-    # r = add_silence(r, 0.5)
-    return sample_width, r
+    sr = shared.setting['sample_rate']
+    # todo